fixed the sync issues between the client and firebase
fixed the sync issues between the client and firebase by refreshing the
client side data for the question before updating the vote count.
fixed issue: "If a user votes on a question. It resets the question to
the wording the User saw even if customer edited since the User loaded
the page."
fixed issue: "If user A load page if that has Question Z. If user B then
loads Question Z and answers yes. Then user A also answers yes then the
yes count of Question Z go up by 1 instead of 2."